+++ lib/src/auth/presentation/containers/info_app/info_app_provider.dart
@@ -1,5 +1,5 @@
+import 'package:flutter_base/src/shared/application/use_cases/get_platform_info_use_dart.dart';
 import 'package:flutter_base/src/shared/domain/models/platform_info.dart';
-import 'package:flutter_base/src/shared/domain/use_cases/get_platform_info_use_dart.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
