+++ lib/src/auth/presentation/pages/forgot_password_confirm_page.dart
@@ -1,7 +1,7 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_base/src/auth/application/forgot_password_confirm_provider.dart';
 import 'package:flutter_base/src/auth/presentation/pages/change_password_page.dart';
+import 'package:flutter_base/src/auth/presentation/states/forgot_password_confirm_provider.dart';
 import 'package:flutter_base/src/shared/presentation/i18n/locale_keys.g.dart';
 import 'package:flutter_base/src/shared/presentation/utils/styles/insets.dart';
 import 'package:flutter_base/src/shared/presentation/widgets/components/box_spacer.dart';
