+++ lib/src/auth/presentation/states/forgot_password_provider.dart
@@ -1,7 +1,7 @@
-import 'package:flutter_base/src/auth/domain/use_cases/reset_password_use_case.dart';
+import 'package:flutter_base/src/auth/application/use_cases/reset_password_use_case.dart';
 import 'package:flutter_base/src/auth/presentation/pages/forgot_password_confirm_page.dart';
 import 'package:flutter_base/src/auth/presentation/view_models/forgot_password_view_model.dart';
-import 'package:flutter_base/src/shared/application/ui_provider.dart';
+import 'package:flutter_base/src/shared/presentation/states/ui_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
 import 'package:go_router/go_router.dart';
