+++ lib/src/auth/presentation/states/login_provider.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_base/src/auth/domain/use_cases/login_use_case.dart';
+import 'package:flutter_base/src/auth/application/use_cases/login_use_case.dart';
 import 'package:flutter_base/src/auth/presentation/view_models/basic_login_view_model.dart';
-import 'package:flutter_base/src/shared/application/ui_provider.dart';
 import 'package:flutter_base/src/shared/domain/models/device_type.dart';
-import 'package:flutter_base/src/user/application/user_provider.dart';
+import 'package:flutter_base/src/shared/presentation/states/ui_provider.dart';
+import 'package:flutter_base/src/user/presentation/states/user_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
 import 'package:go_router/go_router.dart';
