+++ lib/src/auth/presentation/states/social_auth_provider.dart
@@ -1,13 +1,13 @@
 import 'package:easy_localization/easy_localization.dart';
+import 'package:flutter_base/src/auth/application/use_cases/login_use_case.dart';
+import 'package:flutter_base/src/auth/application/use_cases/sign_up_use_case.dart';
 import 'package:flutter_base/src/auth/domain/enums/auth_provider.dart';
-import 'package:flutter_base/src/auth/domain/use_cases/login_use_case.dart';
-import 'package:flutter_base/src/auth/domain/use_cases/sign_up_use_case.dart';
-import 'package:flutter_base/src/shared/application/ui_provider.dart';
 import 'package:flutter_base/src/shared/domain/models/app_error.dart';
 import 'package:flutter_base/src/shared/domain/models/device_type.dart';
 import 'package:flutter_base/src/shared/presentation/i18n/locale_keys.g.dart';
-import 'package:flutter_base/src/user/application/user_provider.dart';
+import 'package:flutter_base/src/shared/presentation/states/ui_provider.dart';
 import 'package:flutter_base/src/user/domain/enums/user_device_type.dart';
+import 'package:flutter_base/src/user/presentation/states/user_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
 import 'package:go_router/go_router.dart';
