+++ lib/src/auth/presentation/states/verify_account_provider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
-import 'package:flutter_base/src/auth/domain/use_cases/verify_account_use_case.dart';
-import 'package:flutter_base/src/shared/application/ui_provider.dart';
-import 'package:flutter_base/src/user/application/user_provider.dart';
+import 'package:flutter_base/src/auth/application/use_cases/verify_account_use_case.dart';
+import 'package:flutter_base/src/shared/presentation/states/ui_provider.dart';
+import 'package:flutter_base/src/user/presentation/states/user_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
