+++ lib/src/user/presentation/states/edit_avatar_provider.dart
@@ -3,14 +3,14 @@ import 'dart:async';
 import 'package:extended_image/extended_image.dart';
 import 'package:flutter/foundation.dart' show kIsWeb, Uint8List;
 import 'package:flutter/widgets.dart';
-import 'package:flutter_base/src/shared/application/ui_provider.dart';
+import 'package:flutter_base/src/shared/application/use_cases/image_from_camera_use_case.dart';
+import 'package:flutter_base/src/shared/application/use_cases/image_from_gallery_use_case.dart';
 import 'package:flutter_base/src/shared/domain/interfaces/i_edit_image_service.dart';
-import 'package:flutter_base/src/shared/domain/use_cases/image_from_camera_use_case.dart';
-import 'package:flutter_base/src/shared/domain/use_cases/image_from_gallery_use_case.dart';
 import 'package:flutter_base/src/shared/presentation/router/app_router.dart';
-import 'package:flutter_base/src/user/application/user_provider.dart';
+import 'package:flutter_base/src/shared/presentation/states/ui_provider.dart';
 import 'package:flutter_base/src/user/domain/interfaces/i_user_repository.dart';
 import 'package:flutter_base/src/user/presentation/pages/edit_avatar_page.dart';
+import 'package:flutter_base/src/user/presentation/states/user_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_it/get_it.dart';
 import 'package:go_router/go_router.dart';
