+++ lib/src/user/presentation/widgets/containers/profile_photo_action_sheet.dart
@@ -1,7 +1,7 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_base/src/shared/presentation/i18n/locale_keys.g.dart';
-import 'package:flutter_base/src/user/application/edit_avatar_provider.dart';
+import 'package:flutter_base/src/user/presentation/states/edit_avatar_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class ProfilePhotoActionSheet extends ConsumerWidget {
