+++ test/src/auth/presentation/pages/forgot_password_confirm_page_test.dart
@@ -4,9 +4,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_base/src/auth/domain/interfaces/i_auth_repository.dart';
 import 'package:flutter_base/src/auth/presentation/pages/change_password_page.dart';
 import 'package:flutter_base/src/auth/presentation/pages/forgot_password_confirm_page.dart';
-import 'package:flutter_base/src/shared/application/deep_link_controller.dart';
 import 'package:flutter_base/src/shared/domain/interfaces/i_deep_link_service.dart';
 import 'package:flutter_base/src/shared/presentation/i18n/locale_keys.g.dart';
+import 'package:flutter_base/src/shared/presentation/states/deep_link_controller.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mocktail/mocktail.dart';
