+++ test/src/auth/presentation/pages/login_page_test.dart
@@ -6,9 +6,9 @@ import 'package:flutter_base/src/auth/domain/models/token_model.dart';
 import 'package:flutter_base/src/notifications/domain/interfaces/i_notifications_service.dart';
 import 'package:flutter_base/src/posts/presentation/pages/post_page.dart';
 import 'package:flutter_base/src/shared/domain/models/app_error.dart';
-import 'package:flutter_base/src/user/application/user_provider.dart';
 import 'package:flutter_base/src/user/domain/interfaces/i_user_repository.dart';
 import 'package:flutter_base/src/user/domain/models/user.dart';
+import 'package:flutter_base/src/user/presentation/states/user_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
