+++ test/src/user/presentation/pages/profile_test.dart
@@ -4,11 +4,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_base/src/auth/domain/interfaces/i_auth_repository.dart';
 import 'package:flutter_base/src/auth/domain/interfaces/i_token_repository.dart';
 import 'package:flutter_base/src/auth/presentation/pages/initial_page.dart';
-import 'package:flutter_base/src/user/application/user_provider.dart';
 import 'package:flutter_base/src/user/domain/interfaces/i_user_repository.dart';
 import 'package:flutter_base/src/user/domain/models/user.dart';
 import 'package:flutter_base/src/user/presentation/pages/edit_profile_page.dart';
 import 'package:flutter_base/src/user/presentation/pages/profile_page.dart';
+import 'package:flutter_base/src/user/presentation/states/user_provider.dart';
 import 'package:flutter_base/src/user/presentation/widgets/containers/profile_photo_action_sheet.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
